+++ cmd/runner/metrics.go
@@ -1,0 +1,130 @@
+// Copyright 2021 (c) Cognizant Digital Business, Evolutionary AI. All rights reserved. Issued under the Apache 2.0 License.
+
+package main
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+	dto "github.com/prometheus/client_model/go"
+
+	"github.com/jjeffery/kv"
+)
+
+var (
+	refreshSuccesses = prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "runner_queue_refresh_success",
+			Help: "Number of successful queue inventory checks.",
+		},
+		[]string{"host", "project"},
+	)
+	refreshFailures = prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "runner_queue_refresh_fail",
+			Help: "Number of failed queue inventory checks.",
+		},
+		[]string{"host", "project"},
+	)
+
+	queueChecked = prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "runner_queue_checked",
+			Help: "Number of times a queue is queried for work.",
+		},
+		[]string{"host", "queue_type", "queue_name"},
+	)
+	queueIgnored = prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "runner_queue_ignored",
+			Help: "Number of times a queue is intentionally not queried, or skipped work.",
+		},
+		[]string{"host", "queue_type", "queue_name"},
+	)
+	queueRunning = prometheus.NewGaugeVec(
+		prometheus.GaugeOpts{
+			Name: "runner_project_running",
+			Help: "Number of experiments being actively worked on per queue.",
+		},
+		[]string{"host", "queue_type", "queue_name", "project", "experiment"},
+	)
+	queueRan = prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "runner_project_completed",
+			Help: "Number of experiments that have been run per queue.",
+		},
+		[]string{"host", "queue_type", "queue_name", "project", "experiment"},
+	)
+)
+
+func init() {
+	prometheus.MustRegister(refreshSuccesses)
+	prometheus.MustRegister(refreshFailures)
+	prometheus.MustRegister(queueChecked)
+	prometheus.MustRegister(queueIgnored)
+	prometheus.MustRegister(queueRunning)
+	prometheus.MustRegister(queueRan)
+}
+
+func GetCounterValue(metric *prometheus.CounterVec, labels prometheus.Labels) (val float64, err kv.Error) {
+	m := &dto.Metric{}
+	if errGo := metric.With(labels).Write(m); errGo != nil {
+		return 0, kv.Wrap(errGo)
+	}
+	return m.Counter.GetValue(), nil
+}
+
+func GetGaugeValue(metric *prometheus.GaugeVec, labels prometheus.Labels) (val float64, err kv.Error) {
+	m := &dto.Metric{}
+	if errGo := metric.With(labels).Write(m); errGo != nil {
+		return 0, kv.Wrap(errGo)
+	}
+	return m.Counter.GetValue(), nil
+}
+
+func getMetricValue(m *dto.Metric) float64 {
+	switch {
+	case m.Gauge != nil:
+		return m.GetGauge().GetValue()
+	case m.Counter != nil:
+		return m.GetCounter().GetValue()
+	case m.Untyped != nil:
+		return m.GetUntyped().GetValue()
+	default:
+		return 0.0
+	}
+}
+
+func GetCounterAccum(counter *prometheus.CounterVec) (val float64, err kv.Error) {
+	mC := make(chan prometheus.Metric, 1)
+
+	go func() {
+		counter.Collect(mC)
+		close(mC)
+	}()
+
+	for metric := range mC {
+		m := &dto.Metric{}
+		if errGo := metric.Write(m); errGo != nil {
+			return 0, kv.Wrap(errGo)
+		}
+		val += getMetricValue(m)
+	}
+	return val, nil
+}
+
+func GetGaugeAccum(gauge *prometheus.GaugeVec) (val float64, err kv.Error) {
+	mC := make(chan prometheus.Metric, 1)
+
+	go func() {
+		gauge.Collect(mC)
+		close(mC)
+	}()
+
+	for metric := range mC {
+		m := &dto.Metric{}
+		if errGo := metric.Write(m); errGo != nil {
+			return 0, kv.Wrap(errGo)
+		}
+		val += getMetricValue(m)
+	}
+	return val, nil
+}