+++ CHANGELOG.md
@@ -302,6 +302,7 @@ FIXES:
 IMPROVEMENTS:
 
 * Storage limitations now used when downloading artifacts, based on the requested disk space from the StudioML client
+* Idle Time and Jobs completed limits added, new options -limit-idle-duration duration, -limit-interval duration, -limit-tasks uint
 
 FIXES: